[refactor] Split utils as crud*
